+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+- Add `AbstractCompilerPass::checkMandatoryServices` to quickly checks if the services listed in th e`MANDATORY_SERVICES` constant exists in the container.
+- Add `AbstractConfiguration::getKey` and `AbstractConfiguration::get` to easily retrieve a configuration. It allows also more traceability of the configuration usage.
+
 ## 1.0.0
 
 - Add `AbstractCompilerPass`